Use NotLoadedCheck when the code cannot be found
