+++ colin/core/checks/filesystem.py
@@ -17,7 +17,7 @@ import logging
 
 from conu.exceptions import ConuException
 
-from .images import ImageAbstractCheck
+from .abstract_check import ImageAbstractCheck
 from ..result import CheckResult
 
 logger = logging.getLogger(__name__)
