+++ colin/core/checks/images.py
@@ -13,13 +13,3 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
-
-from .abstract_check import AbstractCheck
-
-
-class ImageAbstractCheck(AbstractCheck):
-    check_type = "image"
-
-
-class FilesystemAbstractCheck(AbstractCheck):
-    pass
